Improve search in sidebar
